refactor(serverless): isolate Vercel lifecycle adapter
